requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
schedule==1.2.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1 